--- a/Dockerfile
+++ b/Dockerfile
@@ -319,6 +319,12 @@ COPY --from=router-builder /usr/local/cargo/bin/text-generation-router /usr/loca
 # Install launcher
 COPY --from=launcher-builder /usr/local/cargo/bin/text-generation-launcher /usr/local/bin/text-generation-launcher
 
+# Install cufile.so, libnuma.so, and fastsafetensors
+RUN dnf config-manager \
+       --add-repo https://developer.download.nvidia.com/compute/cuda/repos/rhel9/x86_64/cuda-rhel9.repo \
+    && dnf install -y libcufile-12-1 numactl-libs
+RUN pip install -v fastsafetensors==0.1.0 --no-cache-dir
+
 ENV PORT=3000 \
     GRPC_PORT=8033 \
     HOME=/home/tgis