@njhill 2024-05-25
Thanks @takeshi-yoshimura for the great work!

I have a few initial thoughts/questions:

I feel it would be better for the fastsafetensors library to be fully independent and not contain any inference server specific code, i.e. include the code from https://github.com/foundation-model-stack/fastsafetensors/blob/main/fastsafetensors/connectors/tgis_weights.py as a file in this PR.

> Currently, the class has `get_tensor` and `get_sahrded` and others except for quantization.

Does this mean it won't work for quantized models? Is this particularly difficult or just something that you haven't implemented yet?

> However, GDS limits the available optimization using `PYTORCH_CUDA_ALLOC_CONF`. Currently, we cannot set it to "expandable_segments:True" to use cuMemmap, which is not available with peer-to-peer transfers.

This optimization is only beneficial in some cases (for example I don't think its helpful when using paged attention). But in the cases that it is effective, we probably wouldn't want to use the faster loading because runtime performance is more important.

> Additionally, using tensor parallel with Fastsafetensors can cause deadlocks at memory estimation. In my experiments, I always set `ESTIMATE_MEMORY=off` to avoid this issue.

I think this is more of a problem and we'd need to understand and resolve it. Is the GPU memory imbalance you mentioned just during the model loading phase, or it remains the case at runtime (I don't really understand if it's the latter).

> I added two tuning knobs for the number of copy threads and bounce buffer sizes. They should be changed according to the available CPUs in a NUMA node and the size of the CPU cache

Could these be auto-configured? Ideally this would not be something that users have to tune on a case-by-case basis.