@takeshi-yoshimura 2024-05-27
Thanks @njhill.

> I feel it would be better for the fastsafetensors library to be fully independent and not contain any inference server specific code, i.e. include the code from https://github.com/foundation-model-stack/fastsafetensors/blob/main/fastsafetensors/connectors/tgis_weights.py as a file in this PR.

I have just pushed the change including upgrading fastsafetensors to be 0.1.1 for minor bug fixes. I noticed that the Weight class has slightly changed since what fastsafetensors has.

> Does this mean it won't work for quantized models? Is this particularly difficult or just something that you haven't implemented yet?

I just haven't implemented yet and I believe there are no technical difficulties. The Weight class just raises a NotImplementedError to call `get_multi_weights_col` and `get_multi_weights_row` with quantization parameters.

> But in the cases that it is effective, we probably wouldn't want to use the faster loading because runtime performance is more important.

In that case, users can just turn off GDS by setting FST_NOGDS=1. Fastsafetensors still provide optimized file I/O and GPU offloading without GDS.

> I think this is more of a problem and we'd need to understand and resolve it. Is the GPU memory imbalance you mentioned just during the model loading phase, or it remains the case at runtime (I don't really understand if it's the latter).

If I remember correctly, the failure happened after the model loading phase at https://github.com/IBM/text-generation-inference/blob/main/server/text_generation_server/utils/memory_characterizer.py#L360. fastsafetensors returns unused GPU memory to torch's CUDACachingAllocator, so it should not cause the imbalance or memory leaks. But it may cause fragmentations and raise OoM at model loading with TP.

I encountered the problem only when I tried bloom with TP=8. I could not reproduce the issue with llama-13b and falcon-40b on TP=2. I will investigate memory trace later to make sure if this happens on current TGIS and fastsafetensors.

> Could these be auto-configured? Ideally this would not be something that users have to tune on a case-by-case basis.

Yes, we can provide auto-tuning configuration via environmental variables FST_TUNE=auto to set things according to number of CPUs and cache sizes. I will add the feature later.