@takeshi-yoshimura 2024-05-27
I have added an auto-configuration. Thanks.